# internal/transport/http/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "企业 AI Coding 能力治理平台。管理接口先登录获取 data.token，再在 Authorize 中填写 AdminBearer（Bearer + 空格 + Token）。Gateway 使用成员 Access Key，与管理员 Token 分开。JSON 中的业务 ID 使用字符串。",
        "title": "Zentrola API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/auth/setup": {
            "get": {
                "description": "只返回是否尚未创建管理员；已有任何管理员记录均不重新开放初始化。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "查询是否需要初始化管理员",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/security.SetupStatus"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    }
                }
            },
            "post": {
                "description": "仅未初始化时可用；账号 1～64 bytes，密码至少 6 个字符且不超过 72 bytes；创建成功后使用登录接口。已有管理员时返回 ALREADY_INITIALIZED。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "创建系统首位管理员",
                "parameters": [
                    {
                        "description": "初始管理员账号与密码",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.SetupResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/usage": {
            "get": {
                "security": [
                    {
                        "AdminBearer": []
                    }
                ],
                "description": "只查询当前组织；按 ID 倒序分页，最新记录在前；时间区间最长 366 天。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用量统计"
                ],
                "summary": "用量记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "上一页 nextCursor，默认从头查询",
                        "name": "after",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 50,
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "起始时间 RFC3339，默认 to 前 24 小时",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束时间 RFC3339，默认当前时间",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "成员 ID",
                        "name": "memberId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "模型 ID",
                        "name": "modelId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "资源 ID",
                        "name": "resourceId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.PageResponse-usage_Row"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "请求追踪 ID"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    }
                }
            }
        },
        "/api/v1/usage/dashboard": {
            "get": {
                "security": [
                    {
                        "AdminBearer": []
                    }
                ],
                "description": "返回当前组织激活用户数、启用模型和服务商数量，以及指定时间范围内的 Token 汇总、用户 Token Top 10 与模型请求 Top 10。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用量统计"
                ],
                "summary": "首页统计",
                "parameters": [
                    {
                        "type": "string",
                        "description": "起始时间 RFC3339",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "结束时间 RFC3339",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/usage.Dashboard"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    }
                }
            }
        },
        "/api/v1/usage/writer": {
            "get": {
                "security": [
                    {
                        "AdminBearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用量统计"
                ],
                "summary": "用量写入队列指标",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/usage.Metrics"
                                        }
                                    }
                                }
                            ]
                        },
                        "headers": {
                            "X-Request-ID": {
                                "type": "string",
                                "description": "请求追踪 ID"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_zentrola_zentrola_internal_transport_http.response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_zentrola_zentrola_internal_transport_http.response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                }
            }
        },
...
        "http.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "example": "your-admin-password"
                },
                "username": {
                    "type": "string",
                    "example": "admin"
                }
            }
        },
...
        "http.PageResponse-usage_Row": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usage.Row"
                    }
                },
                "nextCursor": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "http.SetupResponse": {
            "type": "object",
            "properties": {
                "initialized": {
                    "type": "boolean"
                }
            }
        },
...
        "security.SetupStatus": {
            "type": "object",
            "properties": {
                "required": {
                    "type": "boolean"
                }
            }
        },
        "usage.Dashboard": {
            "type": "object",
            "properties": {
                "activeMemberCount": {
                    "type": "integer"
                },
                "modelCount": {
                    "type": "integer"
                },
                "modelRanking": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usage.ModelRank"
                    }
                },
                "providerCount": {
                    "type": "integer"
                },
                "tokenRanking": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usage.TokenRank"
                    }
                },
                "totalTokens": {
                    "type": "integer"
                }
            }
        },
        "usage.Metrics": {
            "type": "object",
            "properties": {
                "failed": {
                    "type": "integer"
                },
                "fallback": {
                    "type": "integer"
                },
                "pending": {
                    "type": "integer"
                },
                "persisted": {
                    "type": "integer"
                },
                "queued": {
                    "type": "integer"
                },
                "writeErrors": {
                    "type": "integer"
                }
            }
        },
        "usage.ModelRank": {
            "type": "object",
            "properties": {
                "requests": {
                    "type": "integer"
                },
                "tokens": {
                    "type": "integer"
                },
                "upstreamModelCode": {
                    "type": "string"
                }
            }
        },
        "usage.Row": {
            "type": "object",
            "properties": {
                "attemptNo": {
                    "type": "integer"
                },
                "cachedInputTokens": {
                    "type": "integer"
                },
                "clientProtocol": {
                    "type": "string"
                },
                "completedAt": {
                    "type": "string"
                },
                "errorType": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "0"
                },
                "inputTokens": {
                    "type": "integer"
                },
                "latencyMs": {
                    "type": "integer"
                },
                "modelId": {
                    "type": "string",
                    "example": "0"
                },
                "outputTokens": {
                    "type": "integer"
                },
                "principalId": {
                    "type": "string",
                    "example": "0"
                },
                "principalName": {
                    "type": "string"
                },
                "providerId": {
                    "type": "string",
                    "example": "0"
                },
                "providerModelId": {
                    "type": "string",
                    "example": "0"
                },
                "requestAt": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                },
                "resourceId": {
                    "type": "string",
                    "example": "0"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "usage.TokenRank": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "principalId": {
                    "type": "string",
                    "example": "0"
                },
                "tokens": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminBearer": {
            "description": "管理员登录返回的 Token，填写：Bearer \u003ctoken\u003e。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "GatewayBearer": {
            "description": "成员 Access Key，填写：Bearer \u003caccess-key\u003e。可用于 OpenAI 或 Anthropic 接口。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "GatewayKey": {
            "description": "成员 Access Key 原文，仅用于 Anthropic 接口；不要同时发送 Authorization。",
            "type": "apiKey",
            "name": "x-api-key",
            "in": "header"
        }
    }
}